modified and switch int to double

+++ main.cpp
@@ -4,13 +4,13 @@
 
 int main() 
 {
-    int num1 = getNum();
+    double num1 = getNum();
 
     int gop = getOp();
 
-    int num2 = getNum();
+    double num2 = getNum();
 
-    int result = calcResult(num1, gop, num2);
+    double result = calcResult(num1, gop, num2);
 
     finalResult(result);
